Move utils functions to appropriate objects
