+++ src/utils/index.ts
@@ -2,15 +2,10 @@ import {ConfigInterface} from '../base-types';
 
 import {createDecodeSessionToken} from './decode-session-token';
 import {nonce} from './nonce';
-import {createGraphqlProxy} from './graphql_proxy';
 import {safeCompare} from './safe-compare';
-import {createValidateHmac} from './hmac-validator';
 import {createSanitizeShop, createSanitizeHost} from './shop-validator';
+import {createValidateHmac} from './hmac-validator';
 import {createVersionCompatible} from './version-compatible';
-import {
-  createGetEmbeddedAppUrl,
-  createBuildEmbeddedAppUrl,
-} from './get-embedded-app-url';
 
 // eslint-disable-next-line no-warning-comments
 // TODO refactor utils functions to take in objects for consistency (and update docs)
@@ -18,13 +13,10 @@ export function shopifyUtils(config: ConfigInterface) {
   return {
     decodeSessionToken: createDecodeSessionToken(config),
     nonce,
-    graphqlProxy: createGraphqlProxy(config),
     safeCompare,
-    validateHmac: createValidateHmac(config),
     sanitizeShop: createSanitizeShop(config),
     sanitizeHost: createSanitizeHost(config),
+    validateHmac: createValidateHmac(config),
     versionCompatible: createVersionCompatible(config),
-    getEmbeddedAppUrl: createGetEmbeddedAppUrl(config),
-    buildEmbeddedAppUrl: createBuildEmbeddedAppUrl(config),
   };
 }
